Control.Monad.Foil.TH.MkToFoil

Documentation

mkToFoil

Arguments

:: Name
Type name for raw terms.

-> Name
Type name for raw variable identifiers.

-> Name
Type name for raw scoped terms.

-> Name
Type name for raw patterns.

-> Q [Dec]

Generate conversion functions from raw to scope-safe representation.

Since: 0.0.1

mkExtendScopeFoilPattern

Arguments

:: Name
Type name for raw variable identifiers.

-> Name
Type name for raw patterns.

-> Q [Dec]

Generate a function to extend scope with variables from a given pattern.

Since: 0.0.1

mkWithRefreshedFoilPattern

Arguments

:: Name
Type name for raw variable identifiers.

-> Name
Type name for raw patterns.

-> Q [Dec]

Generate a function to extend scope with variables from a given pattern.

Since: 0.0.1

mkToFoilTerm

Arguments

:: Name
Type name for raw terms.

-> Name
Type name for raw variable identifiers.

-> Name
Type name for raw scoped terms.

-> Name
Type name for raw patterns.

-> Q [Dec]

Generate a conversion function from raw terms to scope-safe terms.

Since: 0.0.1

mkToFoilPattern

Arguments

:: Name
Type name for raw variable identifiers.

-> Name
Type name for raw patterns.

-> Q [Dec]

Generate a conversion function from raw terms to scope-safe terms.

Since: 0.1.0