Control.Monad.Free.Foil.Example

Description

Example implementation of untyped \(\lambda\)-calculus with the foil.

Documentation

>>> import Control.Monad.Foil

data ExprF scope term

Untyped \(\lambda\)-terms in scope n.

Since: 0.0.1

Constructors

AppF term term
Application of one term to another: \((t_1, t_2)\)

LamF scope
\(\lambda\)-abstraction introduces a binder and a term in an extended scope: \(\lambda x. t\)

pattern AppE :: AST binder ExprF n -> AST binder ExprF n -> AST binder ExprF n

Application of one term to another.

Since: 0.0.1

pattern LamE :: () => binder n l -> AST binder ExprF l -> AST binder ExprF n

A \(\lambda\)-abstraction, binding a pattern in a term of the extended scope.

Since: 0.0.1

type Expr = AST NameBinder ExprF

A scope-safe \(\lambda\)-term with one name per binder.

Since: 0.0.1

whnf :: forall (n :: S). Distinct n => Scope n -> Expr n -> Expr n

Compute weak head normal form (WHNF) of a \(\lambda\)-term.

>>> whnf emptyScope (AppE (churchN 2) (churchN 2))
λx1. (λx0. λx1. (x0 (x0 x1)) (λx0. λx1. (x0 (x0 x1)) x1))

Since: 0.0.1

whnf' :: Expr 'VoidS -> Expr 'VoidS

Compute weak head normal form (WHNF) of a closed \(\lambda\)-term.

>>> whnf' (AppE (churchN 2) (churchN 2))
λx1. (λx0. λx1. (x0 (x0 x1)) (λx0. λx1. (x0 (x0 x1)) x1))

Since: 0.0.1

nf :: forall (n :: S). Distinct n => Scope n -> Expr n -> Expr n

Compute normal form (NF) of a \(\lambda\)-term.

>>> nf emptyScope (AppE (churchN 2) (churchN 2))
λx1. λx2. (x1 (x1 (x1 (x1 x2))))

Since: 0.0.1

nf' :: Expr 'VoidS -> Expr 'VoidS

Compute normal form (NF) of a closed \(\lambda\)-term.

>>> nf' (AppE (churchN 2) (churchN 2))
λx1. λx2. (x1 (x1 (x1 (x1 x2))))

Since: 0.0.1

ppName :: forall (n :: S). Name n -> String

Pretty print a name.

Since: 0.0.1

ppExpr :: forall (n :: S). Expr n -> String

Pretty-print a \(\lambda\)-term.

>>> ppExpr (churchN 3)
"\955x0. \955x1. (x0 (x0 (x0 x1)))"

Since: 0.0.1

lam :: forall (n :: S). Distinct n => Scope n -> (forall (l :: S). DExt n l => Scope l -> NameBinder n l -> Expr l) -> Expr n

A helper for constructing \(\lambda\)-abstractions.

Since: 0.0.1

churchN :: Int -> Expr 'VoidS

Church-encoding of a natural number \(n\).

>>> churchN 0
λx0. λx1. x1

>>> churchN 3
λx0. λx1. (x0 (x0 (x0 x1)))

Since: 0.0.1