Data.ZipMatchK

Description

Kind-polymorphic syntactic (first-order) unification.

Documentation

module Data.ZipMatchK.Mappings

class ZipMatchK (f :: k) where

Kind-polymorphic syntactic (first-order) unification of two values.

Note that f is expected to be a traversable n-functor. There is no TraversableK constraint to say so at the moment.

The default implementation is generic, via RepK. It is convenient, but it reflects each node into a representation and back on every comparison, and a constructor is represented as a chain of L1/R1 wrappers as long as its index, so the cost grows with the size of the signature. Comparing terms is most of what a typechecker does, and on a large signature the generic instance costs roughly twice what the written-out one does. Use deriveZipMatchK (or deriveZipMatchK2, for a signature bifunctor with extra parameters) to generate the written-out instance instead.

Since: 0.2.0

Minimal complete definition

Nothing

Methods

zipMatchWithK :: forall (as :: LoT k) (bs :: LoT k) (cs :: LoT k). Mappings as bs cs -> (f :@@: as) -> (f :@@: bs) -> Maybe (f :@@: cs)

Perform one level of equality testing:

when k = Type, values are compared directly (e.g. via Eq);

when k = Type -> Type, term constructors are compared. Unequal constructors give Nothing, and equal ones have all their components paired up with a given function.

Since: 0.2.0

default zipMatchWithK :: forall (as :: LoT k) (bs :: LoT k) (cs :: LoT k). (GenericK f, GZipMatch (RepK f), ReqsZipMatchWith (RepK f) as bs cs) => Mappings as bs cs -> (f :@@: as) -> (f :@@: bs) -> Maybe (f :@@: cs)

zipMatchK :: forall {k} (f :: k) (as :: LoT k) (bs :: LoT k). (ZipMatchK f, PairMappings as bs) => (f :@@: as) -> (f :@@: bs) -> Maybe (f :@@: ZipLoT as bs)

Perform one level of equality testing for two values and pair up components using (,):

zipMatchK = zipMatchWithK (\x y -> Just (,) :^: M0)

Since: 0.2.0

Specializations

Unification of plain Types

zipMatchViaEq :: forall {k} a (as :: LoT k) (bs :: LoT k) (cs :: LoT k). Eq a => Mappings as bs cs -> a -> a -> Maybe a

Unify values via Eq. Can be used as an implementation of zipMatchWithK when k = Type.

Since: 0.2.0

zipMatchViaChooseLeft :: forall {k} (as :: LoT k) (bs :: LoT k) (cs :: LoT k) a. Mappings as bs cs -> a -> a -> Maybe a

Always successfully unify any two values of type a by preferring the left value. Can be used as an implementation of zipMatchWithK when k = Type.

Since: 0.2.0

Unification of Functors

zipMatchWith1 :: (Traversable f, ZipMatchK f) => (a -> a' -> Maybe a'') -> f a -> f a' -> Maybe (f a'')

zipMatchWithK specialised to functors.

Note: Traversable is a morally correct constraint here.

Since: 0.3.0

zipMatch1 :: (Traversable f, ZipMatchK f) => f a -> f a' -> Maybe (f (a, a'))

zipMatchK specialised to functors.

Note: Traversable is a morally correct constraint here.

Since: 0.3.0

Unification of Bifunctors

zipMatchWith2 :: (Bitraversable f, ZipMatchK f) => (a -> a' -> Maybe a'') -> (b -> b' -> Maybe b'') -> f a b -> f a' b' -> Maybe (f a'' b'')

zipMatchWithK specialised to bifunctors.

Note: Bitraversable is a morally correct constraint here.

Since: 0.3.0

zipMatch2 :: (Bitraversable f, ZipMatchK f) => f a b -> f a' b' -> Maybe (f (a, a') (b, b'))

zipMatchK specialised to bifunctors.

Note: Bitraversable is a morally correct constraint here.

Since: 0.3.0