Data.ZipMatchK.Generic

Documentation

class ZipMatchK (f :: k) where

Kind-polymorphic syntactic (first-order) unification of two values.

Note that f is expected to be a traversable n-functor. There is no TraversableK constraint to say so at the moment.

The default implementation is generic, via RepK. It is convenient, but it reflects each node into a representation and back on every comparison, and a constructor is represented as a chain of L1/R1 wrappers as long as its index, so the cost grows with the size of the signature. Comparing terms is most of what a typechecker does, and on a large signature the generic instance costs roughly twice what the written-out one does. Use deriveZipMatchK (or deriveZipMatchK2, for a signature bifunctor with extra parameters) to generate the written-out instance instead.

Since: 0.2.0

Minimal complete definition

Nothing

Methods

zipMatchWithK :: forall (as :: LoT k) (bs :: LoT k) (cs :: LoT k). Mappings as bs cs -> (f :@@: as) -> (f :@@: bs) -> Maybe (f :@@: cs)

Perform one level of equality testing:

when k = Type, values are compared directly (e.g. via Eq);

when k = Type -> Type, term constructors are compared. Unequal constructors give Nothing, and equal ones have all their components paired up with a given function.

Since: 0.2.0

default zipMatchWithK :: forall (as :: LoT k) (bs :: LoT k) (cs :: LoT k). (GenericK f, GZipMatch (RepK f), ReqsZipMatchWith (RepK f) as bs cs) => Mappings as bs cs -> (f :@@: as) -> (f :@@: bs) -> Maybe (f :@@: cs)

genericZipMatchK :: forall {k} (f :: k) (as :: LoT k) (bs :: LoT k). (GenericK f, GZipMatch (RepK f), ReqsZipMatch (RepK f) as bs, PairMappings as bs) => (f :@@: as) -> (f :@@: bs) -> Maybe (f :@@: ZipLoT as bs)

Generic implementation of zipMatchK.

Since: 0.2.0

genericZipMatchWithK :: forall {k} (f :: k) (as :: LoT k) (bs :: LoT k) (cs :: LoT k). (GenericK f, GZipMatch (RepK f), ReqsZipMatchWith (RepK f) as bs cs) => Mappings as bs cs -> (f :@@: as) -> (f :@@: bs) -> Maybe (f :@@: cs)

Generic implementation of zipMatchWithK.

Since: 0.2.0

type ReqsZipMatch (f :: LoT k -> Type) (as :: LoT k) (bs :: LoT k) = ReqsZipMatchWith f as bs (ZipLoT as bs)

What gzipMatchWith requires when the components are paired up.

Since: 0.2.0

class GZipMatch (f :: LoT k -> Type) where

Matching on the Generics.Kind representation of a type, which is what the default ZipMatchK instance goes through.

Since: 0.2.0

Associated Types

type ReqsZipMatchWith (f :: LoT k -> Type) (as :: LoT k) (bs :: LoT k) (cs :: LoT k)

What matching this representation requires of its components.

Since: 0.2.0

Methods

gzipMatchWith :: forall (as :: LoT k) (bs :: LoT k) (cs :: LoT k). ReqsZipMatchWith f as bs cs => Mappings as bs cs -> f as -> f bs -> Maybe (f cs)

Match two representations, pairing up their components with the given functions.

Since: 0.2.0

class ZipMatchFields (t :: Atom d Type) where

Matching a single field of a constructor, by the shape of its type.

Since: 0.2.0

Associated Types

type ReqsZipMatchFieldsWith (t :: Atom d Type) (as :: LoT d) (bs :: LoT d) (cs :: LoT d)

What matching this field requires.

Since: 0.2.0

Methods

zipMatchFieldsWith :: forall (as :: LoT d) (bs :: LoT d) (cs :: LoT d). ReqsZipMatchFieldsWith t as bs cs => Mappings as bs cs -> Field t as -> Field t bs -> Maybe (Field t cs)

Match two fields, pairing up their contents.

Since: 0.2.0

Orphan instances

GenericK (,)

Instance details

Associated Types

type RepK (,)

Instance details

Defined in Data.ZipMatchK.Generic

type RepK (,) = Field (Var0 :: Atom (Type -> Type -> Type) Type) :*: Field (Var1 :: Atom (Type -> Type -> Type) Type)

Methods

fromK :: forall (x :: LoT (Type -> Type -> Type)). ((,) :@@: x) -> RepK (,) x

toK :: forall (x :: LoT (Type -> Type -> Type)). RepK (,) x -> (,) :@@: x

GenericK NonEmpty

Instance details

Associated Types

type RepK NonEmpty

Instance details

Defined in Data.ZipMatchK.Generic

type RepK NonEmpty = Field (Var0 :: Atom (Type -> Type) Type) :*: Field ([] :$: (Var0 :: Atom (Type -> Type) Type))

Methods

fromK :: forall (x :: LoT (Type -> Type)). (NonEmpty :@@: x) -> RepK NonEmpty x

toK :: forall (x :: LoT (Type -> Type)). RepK NonEmpty x -> NonEmpty :@@: x

GenericK ((,) a :: Type -> Type)

Instance details

Associated Types

type RepK ((,) a :: Type -> Type)

Instance details

Defined in Data.ZipMatchK.Generic

type RepK ((,) a :: Type -> Type) = Field ('Kon a :: Atom (Type -> Type) Type) :*: Field (Var0 :: Atom (Type -> Type) Type)

Methods

fromK :: forall (x :: LoT (Type -> Type)). ((,) a :@@: x) -> RepK ((,) a) x

toK :: forall (x :: LoT (Type -> Type)). RepK ((,) a) x -> (,) a :@@: x