Control.Monad.Foil.Internal

Description

Main definitions of the foil that can be reused for specific implementations. This is an internal module, so it also contains implementation details of the foil.

The original description of this approach is described in the IFL 2022 paper by Maclaurin, Radul, and Paszke «The Foil: Capture-Avoiding Substitution With No Sharp Edges». This module also introduces CoSinkable class, generalizing handling of patterns, as described in «Free Foil: Generating Efficient and Scope-Safe Abstract Syntax».

Since the representation of scopes and substitutions is either IntMap or IntSet, many of the operations have a worst-case complexity of \(O(\min(n,W))\). This means that the operation can become linear in the size of the scope \(n\) with a maximum of \(W\) — the number of bits in an Int (32 or 64).

Safe types and operations

data S

S is a data kind of scope indices.

Constructors

VoidS
VoidS is the only explicit scope available to the users, representing an empty scope. All other scopes are represented with type variables, bound in rank-2 polymophic functions like withFreshBinder.

newtype Scope (n :: S)

A safe scope, indexed by a type-level scope index n.

Constructors

UnsafeScope RawScope

newtype Name (n :: S)

A name in a safe scope, indexed by a type-level scope index n.

Constructors

UnsafeName RawName

newtype NameBinder (n :: S) (l :: S)

A name binder is a name that extends scope n to a (larger) scope l.

Constructors

UnsafeNameBinder (Name l)

emptyScope :: Scope 'VoidS

An empty scope (without any names).

extendScope :: forall (n :: S) (l :: S). NameBinder n l -> Scope n -> Scope l

\(O(\min(n,W))\). Extend a scope with one name (safely). Note that as long as the foil is used as intended, the name binder is guaranteed to introduce a name that does not appear in the initial scope.

member :: forall (l :: S) (n :: S). Name l -> Scope n -> Bool

A runtime check for potential name capture.

nameOf :: forall (n :: S) (l :: S). NameBinder n l -> Name l

Extract name from a name binder.

nameId :: forall (l :: S). Name l -> Id

Convert Name into an identifier. This may be useful for printing and debugging.

withFreshBinder :: forall (n :: S) r. Scope n -> (forall (l :: S). NameBinder n l -> r) -> r

Allocate a fresh binder for a given scope.

data DistinctEvidence (n :: S) where

Evidence that scope n contains distinct names.

Constructors

Distinct :: forall (n :: S). Distinct n => DistinctEvidence n

unsafeDistinct :: forall (n :: S). DistinctEvidence n

Unsafely declare that scope n is distinct. Used in unsafeAssertFresh.

data ExtEvidence (n :: S) (l :: S) where

Evidence that scope l extends scope n.

Constructors

Ext :: forall (n :: S) (l :: S). Ext n l => ExtEvidence n l

unsafeExt :: forall (n :: S) (l :: S). ExtEvidence n l

Unsafely declare that scope l extends scope n. Used in unsafeAssertFresh.

withFresh :: forall (n :: S) r. Distinct n => Scope n -> (forall (l :: S). DExt n l => NameBinder n l -> r) -> r

Safely produce a fresh name binder with respect to a given scope.

unsafeAssertFresh :: forall (n :: S) (l :: S) (n' :: S) (l' :: S) r. NameBinder n l -> (DExt n' l' => NameBinder n' l' -> r) -> r

Unsafely declare that a given name (binder) is already fresh in any scope n'.

assertDistinct :: forall (n :: S) (l :: S). Distinct n => NameBinder n l -> DistinctEvidence l

A distinct scope extended with a NameBinder is also distinct.

assertExt :: forall (n :: S) (l :: S). NameBinder n l -> ExtEvidence n l

A distinct scope extended with a NameBinder is also distinct.

withRefreshed

Arguments

:: forall (o :: S) (i :: S) r. Distinct o

=> Scope o
Ambient scope.

-> Name i
Name to refresh (if it clashes with the ambient scope).

-> (forall (o' :: S). DExt o o' => NameBinder o o' -> r)
Continuation, accepting the refreshed name.

-> r

Safely rename (if necessary) a given name to extend a given scope. This is similar to withFresh, except if the name does not clash with the scope, it can be used immediately, without renaming.

unsinkName :: forall (n :: S) (l :: S). NameBinder n l -> Name l -> Maybe (Name n)

Try coercing the name back to the (smaller) scope, given a binder that extends that scope.

Unification of binders

data UnifyNameBinders (n :: S) (l :: S) (r :: S) where

Unification result for two binders, extending some common scope to scopes l and r respectively.

Due to the implementation of the foil,

Constructors

SameNameBinders :: forall (n :: S) (l :: S). UnifyNameBinders n l l
Binders are the same, proving that type parameters l and r are in fact equivalent.

RenameLeftNameBinder :: forall (n :: S) (l :: S) (r :: S). (NameBinder n l -> NameBinder n r) -> UnifyNameBinders n l r
It is possible to safely rename the left binder to match the right one.

RenameRightNameBinder :: forall (n :: S) (r :: S) (l :: S). (NameBinder n r -> NameBinder n l) -> UnifyNameBinders n l r
It is possible to safely rename the right binder to match the left one.

unifyNameBinders :: forall (i :: S) (l :: S) (r :: S). NameBinder i l -> NameBinder i r -> UnifyNameBinders i l r

Unify binders either by asserting that they are the same, or by providing a safe renaming function to convert one binder to another.

Safe sinking

class Sinkable (e :: S -> Type) where

Sinking an expression from scope n into a (usualy extended) scope l, given the renaming (injection from scope n to scope l).

Methods

sinkabilityProof

Arguments

:: forall (n :: S) (l :: S). (Name n -> Name l)
Map names from scope n to a (possibly larger) scope l.

-> e n
Expression with free variables in scope n.

-> e l

An implementation of this method that typechecks proves to the compiler that the expression is indeed Sinkable. However, instead of this implementation, sink should be used at all call sites for efficiency.

sink :: forall e (n :: S) (l :: S). (Sinkable e, DExt n l) => e n -> e l

Efficient version of sinkabilityProof. In fact, once sinkabilityProof typechecks, it is safe to sink by coercion. See Section 3.5 in «The Foil: Capture-Avoiding Substitution With No Sharp Edges» for the details.

extendRenaming

Arguments

:: forall pattern (n :: S) (n' :: S) (l :: S) r. CoSinkable pattern

=> (Name n -> Name n')
Map names from scope n to a (possibly larger) scope n'.

-> pattern n l
A pattern that extends scope n to another scope l.

-> (forall (l' :: S). (Name l -> Name l') -> pattern n' l' -> r)
A continuation, accepting an extended renaming from l to l' (which itself extends n') and a (possibly refreshed) pattern that extends n' to l'.

-> r

Extend renaming when going under a CoSinkable pattern (generalized binder). Note that the scope under pattern is independent of the codomain of the renaming.

This function is used to go under binders when implementing sinkabilityProof and is both a generalization of extendRenamingNameBinder and an efficient implementation of coSinkabilityProof.

extendNameBinderRenaming

Arguments

:: forall pattern (i :: S) (n :: S) (n' :: S) (l :: S) r. CoSinkable pattern

=> (NameBinder i n -> NameBinder i n')
Map names from scope n to a (possibly larger) scope n'.

-> pattern n l
A pattern that extends scope n to another scope l.

-> (forall (l' :: S). (NameBinder n' l -> NameBinder n' l') -> pattern n' l' -> r)
A continuation, accepting an extended renaming from l to l' (which itself extends n') and a (possibly refreshed) pattern that extends n' to l'.

-> r

Extend renaming of binders when going under a CoSinkable pattern (generalized binder). Note that the scope under pattern is independent of the codomain of the renaming.

composeNameBinderRenamings

Arguments

:: forall (n :: S) (i :: S) (i' :: S) (l :: S) (l' :: S). (NameBinder n i -> NameBinder n i')
Rename binders extending scope n from i to i'.

-> (NameBinder i' l -> NameBinder i' l')
Rename binders extending scope i' from l to l'.

-> NameBinder n l

-> NameBinder n l'

Safely compose renamings of name binders. The underlying implementation is

fromNameBinderRenaming :: forall (n :: S) (l :: S) (l' :: S). (NameBinder n l -> NameBinder n l') -> Name l -> Name l'

Convert renaming of name binders into renaming of names in the inner scopes.

extendRenamingNameBinder

Arguments

:: forall (n :: S) (n' :: S) (l :: S) r. (Name n -> Name n')
Map names from scope n to a (possibly larger) scope n'.

-> NameBinder n l
A name binder that extends scope n to another scope l.

-> (forall (l' :: S). (Name l -> Name l') -> NameBinder n' l' -> r)
A continuation, accepting an extended renaming from l to l' (which itself extends n') and a (possibly refreshed) binder that extends n' to l'.

-> r

Extend renaming when going under a NameBinder. Note that the scope under binder is independent of the codomain of the renaming.

Semantically, this function may need to rename the binder (resulting in the new scope l'), to make sure it does not clash with scope n'. However, as it turns out, the foil makes it safe to implement this function as a coercion. See Appendix A in «The Foil: Capture-Avoiding Substitution With No Sharp Edges» for the details.

This function is used to go under binders when implementing sinkabilityProof. A generalization of this function is extendRenaming (which is an efficient version of coSinkabilityProof).

class CoSinkable (pattern :: S -> S -> Type) where

CoSinkable is to patterns (generalized binders) what Sinkable is to expressions.

See Section 2.3 of «Free Foil: Generating Efficient and Scope-Safe Abstract Syntax» for more details.

Methods

coSinkabilityProof

Arguments

:: forall (n :: S) (n' :: S) (l :: S) r. (Name n -> Name n')
Map names from scope n to a (possibly larger) scope n'.

-> pattern n l
A pattern that extends scope n to another scope l.

-> (forall (l' :: S). (Name l -> Name l') -> pattern n' l' -> r)
A continuation, accepting an extended renaming from l to l' (which itself extends n') and a (possibly refreshed) pattern that extends n' to l'.

-> r

An implementation of this method that typechecks proves to the compiler that the pattern is indeed CoSinkable. However, instead of this implementation, extendRenaming should be used at all call sites for efficiency.

Safe substitions

newtype Substitution (e :: S -> Type) (i :: S) (o :: S)

A substitution is a mapping from names in scope i to expressions e o in scope o.

Constructors

UnsafeSubstitution (IntMap (e o))

lookupSubst :: forall e (i :: S) (o :: S). InjectName e => Substitution e i o -> Name i -> e o

Apply substitution to a given name.

identitySubst :: forall (e :: S -> Type) (i :: S). InjectName e => Substitution e i i

Identity substitution maps all names to expresion-variables.

addSubst :: forall e (i :: S) (o :: S) (i' :: S). Substitution e i o -> NameBinder i i' -> e o -> Substitution e i' o

Extend substitution with a particular mapping.

addRename :: forall (e :: S -> Type) (i :: S) (o :: S) (i' :: S). InjectName e => Substitution e i o -> NameBinder i i' -> Name o -> Substitution e i' o

Add variable renaming to a substitution. This includes the performance optimization of eliding names mapped to themselves.

Name maps

newtype NameMap (n :: S) a

A total map from names in scope n to elements of type a.

Constructors

NameMap

Fields

getNameMap :: IntMap a

emptyNameMap :: NameMap 'VoidS a

An empty map belongs in the empty scope.

lookupName :: forall (n :: S) a. Name n -> NameMap n a -> a

Looking up a name should always succeed.

Note that since Name is Sinkable, you can lookup a name from scope n in a NameMap for scope l whenever l extends n.

addNameBinder :: forall (n :: S) (l :: S) a. NameBinder n l -> a -> NameMap n a -> NameMap l a

Extending a map with a single mapping.

Note that the scope parameter of the result differs from the initial map.

Raw types and operations

type Id = Int

We will use Int for efficient representation of identifiers.

type RawName = Id

Raw name is simply an identifier.

type RawScope = IntSet

A raw scope is a set of raw names.

rawFreshName :: RawScope -> RawName

\(O(\min(n, W))\). Generate a fresh raw name that does not appear in a given raw scope.

rawMember :: RawName -> RawScope -> Bool

Check if a raw name is contained in a raw scope.

Constraints

class ExtEndo (n :: S)

Every scope is a (trivial) extension of itself.

Important: this class exists to assist tracking scope extensions for type variables of kind S. Users of the foil are not supposed to implement any instances of ExtEndo.

class (ExtEndo n => ExtEndo l) => Ext (n :: S) (l :: S)

Some scopes are extensions of other scopes.

Important: this class exists to assist tracking scope extensions for type variables of kind S. Users of the foil are not supposed to implement any instances of Ext.

class Distinct (n :: S)

Scopes with distinct names.

Important: this class exists to explicitly mark scopes with distinct names. Users of the foil are not supposed to implement any instances of Distinct.

type DExt (n :: S) (l :: S) = (Distinct l, Ext n l)

Scope extensions with distinct names.

class InjectName (e :: S -> Type) where

Instances of this typeclass possess the ability to inject names. Usually, this is a variable data constructor.

Methods

injectName :: forall (n :: S). Name n -> e n

Inject names into expressions.