Data.ZipMatchK.Mappings

Documentation

type family ZipLoT (as :: LoT k) (bs :: LoT k) :: LoT k where ...

Zip to lists of types into a single list of pair types.

Equations

ZipLoT 'LoT0 'LoT0 = 'LoT0

ZipLoT (a ':&&: as :: LoT (Type -> ks)) (b ':&&: bs :: LoT (Type -> ks)) = (a, b) ':&&: ZipLoT as bs

data Mappings (as :: LoT k) (bs :: LoT k) (cs :: LoT k) where

A collection of zipping functions for zipMatchWithK.

Constructors

M0 :: Mappings 'LoT0 'LoT0 'LoT0
An empty collection (when there no (more) type parameters).

(:^:) infixr 5
A non-empty collection (when there is at least one type parameter).

Fields

:: forall {k1} a b c (as1 :: LoT k1) (bs1 :: LoT k1) (cs1 :: LoT k1). (a -> b -> Maybe c)

Zipping for the first type parameter.

-> Mappings as1 bs1 cs1

Zipping for other type parameters.

-> Mappings (a ':&&: as1) (b ':&&: bs1) (c ':&&: cs1)

class PairMappings (as :: LoT k) (bs :: LoT k) where

Methods

pairMappings :: Mappings as bs (ZipLoT as bs)

A collection of pairing functions (\x y -> Just (x, y)) for zipMatchK.

class ApplyMappings (v :: TyVar d Type) where

Methods

applyMappings

Arguments

:: forall (as :: LoT d) (bs :: LoT d) (cs :: LoT d). Mappings as bs cs
A collection of zipping functions.

-> Interpret ('Var v) as
First collection of values (one per type parameter).

-> Interpret ('Var v) bs
Second collection of values (one per type parameter).

-> Maybe (Interpret ('Var v) cs)

Apply a collection of zipping functions to collections of values.

pairA :: Applicative f => a -> b -> f (a, b)

Pair two values in a context.